Create SECURITY.md
This is a very important thing when creating a project since it permits to know what to do in case of a security issue. This file describes briefly in detail what are the actions to do and what can be done when finding a security issue. Also one can know how much we care about one's security.